+++ lib/clipboard/mac.rb
@@ -8,11 +8,11 @@ module Clipboard
     include Implementation
     extend self
 
-    def paste(_ = nil)
+    def paste(_ = nil, **)
       `pbpaste`
     end
 
-    def copy(data)
+    def copy(data, **)
       Utils.popen "pbcopy", data
 
       true